@@ -1,10 +1,8 @@
-/* eslint-disable */
-// https://github.com/michael-ciniawsky/postcss-load-config
+import autoprefixer from 'autoprefixer'
 
-module.exports = {
+export default {
   plugins: [
-    // https://github.com/postcss/autoprefixer
-    require('autoprefixer')({
+    autoprefixer({
       overrideBrowserslist: [
         'last 4 Chrome versions',
         'last 4 Firefox versions',
@@ -13,15 +11,8 @@ module.exports = {
         'last 4 Android versions',
         'last 4 ChromeAndroid versions',
         'last 4 FirefoxAndroid versions',
-        'last 4 iOS versions'
-      ]
-    })
-
-    // https://github.com/elchininet/postcss-rtlcss
-    // If you want to support RTL css, then
-    // 1. yarn/npm install postcss-rtlcss
-    // 2. optionally set quasar.config.js > framework > lang to an RTL language
-    // 3. uncomment the following line:
-    // require('postcss-rtlcss')
-  ]
+        'last 4 iOS versions',
+      ],
+    }),
+  ],
 }
